Magnus is lazy and so am I

+++ frontend/app/src/main/java/com/example/weshussy/ui/activities/HomeActivity.kt
@@ -161,7 +161,7 @@ fun HomeView(viewModel: HomeViewModel) {
             snackbarHostState.showSnackbar(
                 message = "Notification sent!",
                 actionLabel = "OK",
-                duration = SnackbarDuration.Short
+                duration = SnackbarDuration.Long
             )
             showSnackbar = false
         }
